code/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
numpy==1.23.2
    # via -r requirements.in
pygame==2.1.2
    # via -r requirements.in
pyopengl==3.1.6
    # via -r requirements.in
